move antlr4 query factory to its own file
