+++ core/modules/ccontrol/A4UserQueryFactory.h
@@ -1,0 +1,41 @@
+// -*- LSST-C++ -*-
+/*
+ * LSST Data Management System
+ * Copyright 2018 AURA/LSST.
+ *
+ * This product includes software developed by the
+ * LSST Project (http://www.lsst.org/).
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the LSST License Statement and
+ * the GNU General Public License along with this program.  If not,
+ * see <http://www.lsstcorp.org/LegalNotices/>.
+ */
+
+#ifndef A4USERQUERYFACTORY_H
+#define A4USERQUERYFACTORY_H
+
+#include <memory>
+
+namespace lsst {
+namespace qserv {
+namespace query {
+class SelectStmt;
+}
+
+namespace ccontrol {
+
+std::shared_ptr<query::SelectStmt> a4NewUserQuery(const std::string& userQuery);
+
+}}}
+
+#endif
